legacy stash -p: respect the add.interactive.usebuiltin setting
As `git add` traditionally did not expose the `--patch=<mode>` modes via
command-line options, the scripted version of `git stash` had to call
`git add--interactive` directly.

But this prevents the built-in `add -p` from kicking in, as
`add--interactive` is the Perl script.

So let's introduce support for an optional `<mode>` argument in `git add
--patch[=<mode>]`, and use that in the scripted version of `git stash
-p`, so that the built-in interactive add can do its job if configured.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>